Solution to Check if a Number is Square of Another

+++ Check if a Number is Square of Another@@6788dd63490637cfc44e1ba4/code.c
@@ -3,7 +3,7 @@
 int main(){
     int a, b;
     scanf("%d%d", &a, &b);
-    int sq = sqrt(a);
+    float sq = sqrt(a);
     if (sq==b) printf("Yes");
     else printf("No");
     return 0;
